+++ src/Plugin.php
@@ -114,7 +114,7 @@ class Plugin extends BasePlugin
     private function _bindEvent()
     {
         // Keep a layout save counter -- needed for cases where Field Labels POSTs labels for more than one layout,
-        // such as Craft Commerce
+        // such as Craft Commerce and Verbb Wishlist
         $savedLayouts = 0;
 
         Event::on(Fields::class, Fields::EVENT_AFTER_SAVE_FIELD_LAYOUT, function(Event $event) use(&$savedLayouts) {
@@ -298,7 +298,7 @@ class Plugin extends BasePlugin
         }
 
         // Plugin support
-        foreach (['commerce', 'calendar', 'events', 'gift-voucher'] as $pluginHandle) {
+        foreach (['commerce', 'calendar', 'events', 'gift-voucher', 'wishlist'] as $pluginHandle) {
             if ($pluginsService->isPluginInstalled($pluginHandle) && $pluginsService->isPluginEnabled($pluginHandle)) {
                 $method = '_get' . StringHelper::toCamelCase($pluginHandle) . 'Layouts';
                 $layouts = array_merge($layouts, $this->$method());
@@ -370,6 +370,26 @@ class Plugin extends BasePlugin
         ];
     }
 
+    private function _getWishlistLayouts(): array
+    {
+        $wishlistPlugin = Craft::$app->getPlugins()->getPlugin('wishlist');
+        $wishlistTypes = $wishlistPlugin->getListTypes()->getAllListTypes();
+
+        // List items have their own pages, so we need to map the lists to item layout IDs for use on those pages
+        $wishlistItemsService = $wishlistPlugin->getItems();
+        $wishlistItems = [];
+        $wishlistLists = \verbb\wishlist\elements\ListElement::find()->all();
+
+        foreach ($wishlistLists as $list) {
+            $wishlistItems[(int)$list->id] = $list->getType()->itemFieldLayoutId;
+        }
+
+        return [
+            'wishlistListTypes' => $this->_mapWishlistLayouts($wishlistTypes),
+            'wishlistListItems' => $wishlistItems,
+        ];
+    }
+
     private function _mapLayouts($list): array
     {
         $output = [];
@@ -389,6 +409,20 @@ class Plugin extends BasePlugin
             $output[(int)$item->id] = [
                 'default' => (int)$item->fieldLayoutId,
                 'variant' => (int)$item->variantFieldLayoutId,
+            ];
+        }
+
+        return $output;
+    }
+
+    private function _mapWishlistLayouts($list): array
+    {
+        $output = [];
+
+        foreach ($list as $item) {
+            $output[(int)$item->id] = [
+                'default' => (int)$item->fieldLayoutId,
+                'item' => (int)$item->itemFieldLayoutId,
             ];
         }